Move type registry from identity.js to type-registry.js
