+++ lib/solid/profile.js
@@ -6,6 +6,7 @@ module.exports = SolidProfile
 var rdf = require('../util/rdf-parser').rdflib
 var vocab = require('../vocab')
 var identity = require('../identity')
+var typeRegistry = require('../type-registry')
 var graphUtil = require('../util/graph-util')
 var parseLinks = graphUtil.parseLinks
 
@@ -244,12 +245,12 @@ SolidProfile.prototype.loadTypeRegistry = function loadTypeRegistry (options) {
  */
 SolidProfile.prototype.addTypeRegistry = function addTypeRegistry (graph) {
   // Is this a public type registry?
-  if (identity.isPublicTypeIndex(graph)) {
+  if (typeRegistry.isPublicTypeIndex(graph)) {
     if (!this.typeIndexPublic.graph) {  // only initialize once
       this.typeIndexPublic.uri = graph.uri
       this.typeIndexPublic.graph = graph
     }
-  } else if (identity.isPrivateTypeIndex(graph)) {
+  } else if (typeRegistry.isPrivateTypeIndex(graph)) {
     if (!this.typeIndexPrivate.graph) {
       this.typeIndexPrivate.uri = graph.uri
       this.typeIndexPrivate.graph = graph
